[CC-10858] CR fix: reworked tests.
