+++ .gitignore
@@ -24,5 +24,5 @@
 .grunt
 .sass-cache
 /node_modules/
-/tests/_log/*
-!/tests/_log/.gitkeep
+/tests/_output/*
+!/tests/_output/.gitkeep
