+++ src/SprykerEco/Shared/FactFinderSdk/FactFinderSdkConstants.php
@@ -9,6 +9,8 @@ namespace SprykerEco\Shared\FactFinderSdk;
 
 /**
  * Declares global environment configuration keys. Do not use it for other class constants.
+ *
+ * @SuppressWarnings(ModuleConstantsFormingConstantValuesRule)
  */
 interface FactFinderSdkConstants
 {
