+++ src/SprykerEco/Zed/FactFinderSdk/Dependency/Facade/FactFinderSdkToCurrencyBridge.php
@@ -7,6 +7,9 @@
 
 namespace SprykerEco\Zed\FactFinderSdk\Dependency\Facade;
 
+/**
+ * @SuppressWarnings(BridgeNameRule)
+ */
 class FactFinderSdkToCurrencyBridge implements FactFinderSdkToCurrencyInterface
 {
     /**
