+++ src/SprykerEco/Zed/FactFinderSdk/Dependency/Facade/FactFinderSdkToStoreBridge.php
@@ -7,6 +7,9 @@
 
 namespace SprykerEco\Zed\FactFinderSdk\Dependency\Facade;
 
+/**
+ * @SuppressWarnings(BridgeNameRule)
+ */
 class FactFinderSdkToStoreBridge implements FactFinderSdkToStoreInterface
 {
     /**
