--- tests/Functional.suite.yml
@@ -1,4 +1,0 @@
-class_name: FunctionalTester
-modules:
-    enabled:
-        - Filesystem
