+++ tests/SprykerEcoTest/Client/FactFinderSdk/_support/FactFinderSdkClientTester.php
@@ -1,0 +1,31 @@
+<?php
+
+/**
+ * MIT License
+ * Use of this software requires acceptance of the Evaluation License Agreement. See LICENSE file.
+ */
+
+namespace SprykerEcoTest\Client\FactFinderSdk;
+
+use Codeception\Actor;
+
+/**
+ * Inherited Methods
+ *
+ * @method void wantToTest($text)
+ * @method void wantTo($text)
+ * @method void execute($callable)
+ * @method void expectTo($prediction)
+ * @method void expect($prediction)
+ * @method void amGoingTo($argumentation)
+ * @method void am($role)
+ * @method void lookForwardTo($achieveValue)
+ * @method void comment($description)
+ * @method void pause()
+ *
+ * @SuppressWarnings(PHPMD)
+ */
+class FactFinderSdkClientTester extends Actor
+{
+    use _generated\FactFinderSdkClientTesterActions;
+}
