+++ tests/SprykerEcoTest/Client/FactFinderSdk/_support/_generated/FactFinderSdkClientTesterActions.php
@@ -1,0 +1,16 @@
+<?php  //[STAMP] 7c2320439af8949a1a653dfcd8ff077a
+namespace SprykerEcoTest\Client\FactFinderSdk\_generated;
+
+// This class was automatically generated by build task
+// You should not change it manually as it will be overwritten on next build
+// @codingStandardsIgnoreFile
+
+trait FactFinderSdkClientTesterActions
+{
+    /**
+     * @return \Codeception\Scenario
+     */
+    abstract protected function getScenario();
+
+    
+}
