+++ tests/SprykerEcoTest/Zed/FactFinderSdk/_support/FactFinderSdkZedTester.php
@@ -1,0 +1,35 @@
+<?php
+
+/**
+ * MIT License
+ * Use of this software requires acceptance of the Evaluation License Agreement. See LICENSE file.
+ */
+
+namespace SprykerEcoTest\Zed\FactFinderSdk;
+
+use Codeception\Actor;
+
+/**
+ * Inherited Methods
+ *
+ * @method void wantToTest($text)
+ * @method void wantTo($text)
+ * @method void execute($callable)
+ * @method void expectTo($prediction)
+ * @method void expect($prediction)
+ * @method void amGoingTo($argumentation)
+ * @method void am($role)
+ * @method void lookForwardTo($achieveValue)
+ * @method void comment($description)
+ * @method void pause()
+ *
+ * @SuppressWarnings(PHPMD)
+ */
+class FactFinderSdkZedTester extends Actor
+{
+    use _generated\FactFinderSdkZedTesterActions;
+
+    /**
+     * Define custom actions here
+     */
+}
