+++ tests/SprykerEcoTest/Zed/FactFinderSdk/_support/_generated/FactFinderSdkZedTesterActions.php
@@ -1,0 +1,16 @@
+<?php  //[STAMP] 7ec962eb2d1b68d91ff267bfb661d48e
+namespace SprykerEcoTest\Zed\FactFinderSdk\_generated;
+
+// This class was automatically generated by build task
+// You should not change it manually as it will be overwritten on next build
+// @codingStandardsIgnoreFile
+
+trait FactFinderSdkZedTesterActions
+{
+    /**
+     * @return \Codeception\Scenario
+     */
+    abstract protected function getScenario();
+
+    
+}
