--- tests/Unit.suite.yml
@@ -1,5 +1,0 @@
-class_name: UnitTester
-modules:
-    enabled:
-        - Asserts
-
