--- tests/_bootstrap.php
@@ -1,10 +1,0 @@
-<?php
-
-/**
- * (c) Spryker Systems GmbH copyright protected
- */
-
-use Spryker\Shared\Testify\SystemUnderTestBootstrap;
-
-$bootstrap = SystemUnderTestBootstrap::getInstance();
-$bootstrap->bootstrap(SystemUnderTestBootstrap::APPLICATION_ZED);
