--- tests/_support/Module/Functional.php
@@ -1,73 +1,0 @@
-<?php
-
-/**
- * MIT License
- * Use of this software requires acceptance of the Evaluation License Agreement. See LICENSE file.
- */
-
-namespace FactFinder\Module;
-
-use Codeception\Module;
-use Codeception\TestCase;
-use Propel\Runtime\Propel;
-use Silex\Application;
-use Spryker\Zed\Propel\Communication\Plugin\ServiceProvider\PropelServiceProvider;
-
-class Functional extends Module
-{
-    /**
-     * @param array|null $config
-     */
-    public function __construct($config = null)
-    {
-        parent::__construct($config);
-
-        $propelServiceProvider = new PropelServiceProvider();
-        $propelServiceProvider->boot(new Application());
-    }
-
-    /**
-     * @param \Codeception\TestCase $test
-     *
-     * @return void
-     */
-    public function _before(TestCase $test)
-    {
-        parent::_before($test);
-
-        Propel::getWriteConnection('zed')->beginTransaction();
-    }
-
-    /**
-     * @param \Codeception\TestCase $test
-     *
-     * @return void
-     */
-    public function _after(TestCase $test)
-    {
-        parent::_after($test);
-
-        Propel::getWriteConnection('zed')->rollBack();
-
-        if (session_status() === PHP_SESSION_ACTIVE) {
-            session_destroy();
-        }
-    }
-
-    /**
-     * @param \Codeception\TestCase $test
-     * @param bool $fail
-     *
-     * @return void
-     */
-    public function _failed(TestCase $test, $fail)
-    {
-        parent::_failed($test, $fail);
-
-        Propel::getWriteConnection('zed')->rollBack();
-
-        if (session_status() === PHP_SESSION_ACTIVE) {
-            session_destroy();
-        }
-    }
-}
