--- tests/_support/Module/Unit.php
@@ -1,14 +1,0 @@
-<?php
-
-/**
- * MIT License
- * Use of this software requires acceptance of the Evaluation License Agreement. See LICENSE file.
- */
-
-namespace FactFinder\Module;
-
-use Codeception\Module;
-
-class Unit extends Module
-{
-}
