--- tests/_support/_generated/UnitTesterActions.php
@@ -1,609 +1,0 @@
-<?php  //[STAMP] bae170c2703e3123bcf941bf91ed106e
-namespace FactFinder\_generated;
-
-// This class was automatically generated by build task
-// You should not change it manually as it will be overwritten on next build
-// @codingStandardsIgnoreFile
-
-trait UnitTesterActions
-{
-    /**
-     * @return \Codeception\Scenario
-     */
-    abstract protected function getScenario();
-
-    
-    /**
-     * [!] Method is generated. Documentation taken from corresponding module.
-     *
-     * Checks that two variables are equal. If you're comparing floating-point values,
-     * you can specify the optional "delta" parameter which dictates how great of a precision
-     * error are you willing to tolerate in order to consider the two values equal.
-     *
-     * Regular example:
-     * ```php
-     * <?php
-     * $I->assertEquals(5, $element->getChildrenCount());
-     * ```
-     *
-     * Floating-point example:
-     * ```php
-     * <?php
-     * $I->assertEquals(0.3, $calculator->add(0.1, 0.2), 'Calculator should add the two numbers correctly.', 0.01);
-     * ```
-     *
-     * @param        $expected
-     * @param        $actual
-     * @param string $message
-     * @param float  $delta
-     * @see \Codeception\Module\Asserts::assertEquals()
-     */
-    public function assertEquals($expected, $actual, $message = null, $delta = null) {
-        return $this->getScenario()->runStep(new \Codeception\Step\Action('assertEquals', func_get_args()));
-    }
-
- 
-    /**
-     * [!] Method is generated. Documentation taken from corresponding module.
-     *
-     * Checks that two variables are not equal. If you're comparing floating-point values,
-     * you can specify the optional "delta" parameter which dictates how great of a precision
-     * error are you willing to tolerate in order to consider the two values not equal.
-     *
-     * Regular example:
-     * ```php
-     * <?php
-     * $I->assertNotEquals(0, $element->getChildrenCount());
-     * ```
-     *
-     * Floating-point example:
-     * ```php
-     * <?php
-     * $I->assertNotEquals(0.4, $calculator->add(0.1, 0.2), 'Calculator should add the two numbers correctly.', 0.01);
-     * ```
-     *
-     * @param        $expected
-     * @param        $actual
-     * @param string $message
-     * @param float  $delta
-     * @see \Codeception\Module\Asserts::assertNotEquals()
-     */
-    public function assertNotEquals($expected, $actual, $message = null, $delta = null) {
-        return $this->getScenario()->runStep(new \Codeception\Step\Action('assertNotEquals', func_get_args()));
-    }
-
- 
-    /**
-     * [!] Method is generated. Documentation taken from corresponding module.
-     *
-     * Checks that two variables are same
-     *
-     * @param        $expected
-     * @param        $actual
-     * @param string $message
-     * @see \Codeception\Module\Asserts::assertSame()
-     */
-    public function assertSame($expected, $actual, $message = null) {
-        return $this->getScenario()->runStep(new \Codeception\Step\Action('assertSame', func_get_args()));
-    }
-
- 
-    /**
-     * [!] Method is generated. Documentation taken from corresponding module.
-     *
-     * Checks that two variables are not same
-     *
-     * @param        $expected
-     * @param        $actual
-     * @param string $message
-     * @see \Codeception\Module\Asserts::assertNotSame()
-     */
-    public function assertNotSame($expected, $actual, $message = null) {
-        return $this->getScenario()->runStep(new \Codeception\Step\Action('assertNotSame', func_get_args()));
-    }
-
- 
-    /**
-     * [!] Method is generated. Documentation taken from corresponding module.
-     *
-     * Checks that actual is greater than expected
-     *
-     * @param        $expected
-     * @param        $actual
-     * @param string $message
-     * @see \Codeception\Module\Asserts::assertGreaterThan()
-     */
-    public function assertGreaterThan($expected, $actual, $message = null) {
-        return $this->getScenario()->runStep(new \Codeception\Step\Action('assertGreaterThan', func_get_args()));
-    }
-
- 
-    /**
-     * [!] Method is generated. Documentation taken from corresponding module.
-     *
-     * Checks that actual is greater or equal than expected
-     *
-     * @param        $expected
-     * @param        $actual
-     * @param string $message
-     * @see \Codeception\Module\Asserts::assertGreaterThanOrEqual()
-     */
-    public function assertGreaterThanOrEqual($expected, $actual, $message = null) {
-        return $this->getScenario()->runStep(new \Codeception\Step\Action('assertGreaterThanOrEqual', func_get_args()));
-    }
-
- 
-    /**
-     * [!] Method is generated. Documentation taken from corresponding module.
-     *
-     * Checks that actual is less than expected
-     *
-     * @param        $expected
-     * @param        $actual
-     * @param string $message
-     * @see \Codeception\Module\Asserts::assertLessThan()
-     */
-    public function assertLessThan($expected, $actual, $message = null) {
-        return $this->getScenario()->runStep(new \Codeception\Step\Action('assertLessThan', func_get_args()));
-    }
-
- 
-    /**
-     * [!] Method is generated. Documentation taken from corresponding module.
-     *
-     * Checks that actual is less or equal than expected
-     *
-     * @param        $expected
-     * @param        $actual
-     * @param string $message
-     * @see \Codeception\Module\Asserts::assertLessThanOrEqual()
-     */
-    public function assertLessThanOrEqual($expected, $actual, $message = null) {
-        return $this->getScenario()->runStep(new \Codeception\Step\Action('assertLessThanOrEqual', func_get_args()));
-    }
-
- 
-    /**
-     * [!] Method is generated. Documentation taken from corresponding module.
-     *
-     * Checks that haystack contains needle
-     *
-     * @param        $needle
-     * @param        $haystack
-     * @param string $message
-     * @see \Codeception\Module\Asserts::assertContains()
-     */
-    public function assertContains($needle, $haystack, $message = null) {
-        return $this->getScenario()->runStep(new \Codeception\Step\Action('assertContains', func_get_args()));
-    }
-
- 
-    /**
-     * [!] Method is generated. Documentation taken from corresponding module.
-     *
-     * Checks that haystack doesn't contain needle.
-     *
-     * @param        $needle
-     * @param        $haystack
-     * @param string $message
-     * @see \Codeception\Module\Asserts::assertNotContains()
-     */
-    public function assertNotContains($needle, $haystack, $message = null) {
-        return $this->getScenario()->runStep(new \Codeception\Step\Action('assertNotContains', func_get_args()));
-    }
-
- 
-    /**
-     * [!] Method is generated. Documentation taken from corresponding module.
-     *
-     * Checks that string match with pattern
-     *
-     * @param string $pattern
-     * @param string $string
-     * @param string $message
-     * @see \Codeception\Module\Asserts::assertRegExp()
-     */
-    public function assertRegExp($pattern, $string, $message = null) {
-        return $this->getScenario()->runStep(new \Codeception\Step\Action('assertRegExp', func_get_args()));
-    }
-
- 
-    /**
-     * [!] Method is generated. Documentation taken from corresponding module.
-     *
-     * Checks that string not match with pattern
-     *
-     * @param string $pattern
-     * @param string $string
-     * @param string $message
-     * @see \Codeception\Module\Asserts::assertNotRegExp()
-     */
-    public function assertNotRegExp($pattern, $string, $message = null) {
-        return $this->getScenario()->runStep(new \Codeception\Step\Action('assertNotRegExp', func_get_args()));
-    }
-
- 
-    /**
-     * [!] Method is generated. Documentation taken from corresponding module.
-     *
-     * Checks that a string starts with the given prefix.
-     *
-     * @param string $prefix
-     * @param string $string
-     * @param string $message
-     * @see \Codeception\Module\Asserts::assertStringStartsWith()
-     */
-    public function assertStringStartsWith($prefix, $string, $message = null) {
-        return $this->getScenario()->runStep(new \Codeception\Step\Action('assertStringStartsWith', func_get_args()));
-    }
-
- 
-    /**
-     * [!] Method is generated. Documentation taken from corresponding module.
-     *
-     * Checks that a string doesn't start with the given prefix.
-     *
-     * @param string $prefix
-     * @param string $string
-     * @param string $message
-     * @see \Codeception\Module\Asserts::assertStringStartsNotWith()
-     */
-    public function assertStringStartsNotWith($prefix, $string, $message = null) {
-        return $this->getScenario()->runStep(new \Codeception\Step\Action('assertStringStartsNotWith', func_get_args()));
-    }
-
- 
-    /**
-     * [!] Method is generated. Documentation taken from corresponding module.
-     *
-     * Checks that variable is empty.
-     *
-     * @param        $actual
-     * @param string $message
-     * @see \Codeception\Module\Asserts::assertEmpty()
-     */
-    public function assertEmpty($actual, $message = null) {
-        return $this->getScenario()->runStep(new \Codeception\Step\Action('assertEmpty', func_get_args()));
-    }
-
- 
-    /**
-     * [!] Method is generated. Documentation taken from corresponding module.
-     *
-     * Checks that variable is not empty.
-     *
-     * @param        $actual
-     * @param string $message
-     * @see \Codeception\Module\Asserts::assertNotEmpty()
-     */
-    public function assertNotEmpty($actual, $message = null) {
-        return $this->getScenario()->runStep(new \Codeception\Step\Action('assertNotEmpty', func_get_args()));
-    }
-
- 
-    /**
-     * [!] Method is generated. Documentation taken from corresponding module.
-     *
-     * Checks that variable is NULL
-     *
-     * @param        $actual
-     * @param string $message
-     * @see \Codeception\Module\Asserts::assertNull()
-     */
-    public function assertNull($actual, $message = null) {
-        return $this->getScenario()->runStep(new \Codeception\Step\Action('assertNull', func_get_args()));
-    }
-
- 
-    /**
-     * [!] Method is generated. Documentation taken from corresponding module.
-     *
-     * Checks that variable is not NULL
-     *
-     * @param        $actual
-     * @param string $message
-     * @see \Codeception\Module\Asserts::assertNotNull()
-     */
-    public function assertNotNull($actual, $message = null) {
-        return $this->getScenario()->runStep(new \Codeception\Step\Action('assertNotNull', func_get_args()));
-    }
-
- 
-    /**
-     * [!] Method is generated. Documentation taken from corresponding module.
-     *
-     * Checks that condition is positive.
-     *
-     * @param        $condition
-     * @param string $message
-     * @see \Codeception\Module\Asserts::assertTrue()
-     */
-    public function assertTrue($condition, $message = null) {
-        return $this->getScenario()->runStep(new \Codeception\Step\Action('assertTrue', func_get_args()));
-    }
-
- 
-    /**
-     * [!] Method is generated. Documentation taken from corresponding module.
-     *
-     * Checks that the condition is NOT true (everything but true)
-     *
-     * @param        $condition
-     * @param string $message
-     * @see \Codeception\Module\Asserts::assertNotTrue()
-     */
-    public function assertNotTrue($condition, $message = null) {
-        return $this->getScenario()->runStep(new \Codeception\Step\Action('assertNotTrue', func_get_args()));
-    }
-
- 
-    /**
-     * [!] Method is generated. Documentation taken from corresponding module.
-     *
-     * Checks that condition is negative.
-     *
-     * @param        $condition
-     * @param string $message
-     * @see \Codeception\Module\Asserts::assertFalse()
-     */
-    public function assertFalse($condition, $message = null) {
-        return $this->getScenario()->runStep(new \Codeception\Step\Action('assertFalse', func_get_args()));
-    }
-
- 
-    /**
-     * [!] Method is generated. Documentation taken from corresponding module.
-     *
-     * Checks that the condition is NOT false (everything but false)
-     *
-     * @param        $condition
-     * @param string $message
-     * @see \Codeception\Module\Asserts::assertNotFalse()
-     */
-    public function assertNotFalse($condition, $message = null) {
-        return $this->getScenario()->runStep(new \Codeception\Step\Action('assertNotFalse', func_get_args()));
-    }
-
- 
-    /**
-     * [!] Method is generated. Documentation taken from corresponding module.
-     *
-     * Checks if file exists
-     *
-     * @param string $filename
-     * @param string $message
-     * @see \Codeception\Module\Asserts::assertFileExists()
-     */
-    public function assertFileExists($filename, $message = null) {
-        return $this->getScenario()->runStep(new \Codeception\Step\Action('assertFileExists', func_get_args()));
-    }
-
- 
-    /**
-     * [!] Method is generated. Documentation taken from corresponding module.
-     *
-     * Checks if file doesn't exist
-     *
-     * @param string $filename
-     * @param string $message
-     * @see \Codeception\Module\Asserts::assertFileNotExists()
-     */
-    public function assertFileNotExists($filename, $message = null) {
-        return $this->getScenario()->runStep(new \Codeception\Step\Action('assertFileNotExists', func_get_args()));
-    }
-
- 
-    /**
-     * [!] Method is generated. Documentation taken from corresponding module.
-     *
-     * @param $expected
-     * @param $actual
-     * @param $description
-     * @see \Codeception\Module\Asserts::assertGreaterOrEquals()
-     */
-    public function assertGreaterOrEquals($expected, $actual, $description = null) {
-        return $this->getScenario()->runStep(new \Codeception\Step\Action('assertGreaterOrEquals', func_get_args()));
-    }
-
- 
-    /**
-     * [!] Method is generated. Documentation taken from corresponding module.
-     *
-     * @param $expected
-     * @param $actual
-     * @param $description
-     * @see \Codeception\Module\Asserts::assertLessOrEquals()
-     */
-    public function assertLessOrEquals($expected, $actual, $description = null) {
-        return $this->getScenario()->runStep(new \Codeception\Step\Action('assertLessOrEquals', func_get_args()));
-    }
-
- 
-    /**
-     * [!] Method is generated. Documentation taken from corresponding module.
-     *
-     * @param $actual
-     * @param $description
-     * @see \Codeception\Module\Asserts::assertIsEmpty()
-     */
-    public function assertIsEmpty($actual, $description = null) {
-        return $this->getScenario()->runStep(new \Codeception\Step\Action('assertIsEmpty', func_get_args()));
-    }
-
- 
-    /**
-     * [!] Method is generated. Documentation taken from corresponding module.
-     *
-     * @param $key
-     * @param $actual
-     * @param $description
-     * @see \Codeception\Module\Asserts::assertArrayHasKey()
-     */
-    public function assertArrayHasKey($key, $actual, $description = null) {
-        return $this->getScenario()->runStep(new \Codeception\Step\Action('assertArrayHasKey', func_get_args()));
-    }
-
- 
-    /**
-     * [!] Method is generated. Documentation taken from corresponding module.
-     *
-     * @param $key
-     * @param $actual
-     * @param $description
-     * @see \Codeception\Module\Asserts::assertArrayNotHasKey()
-     */
-    public function assertArrayNotHasKey($key, $actual, $description = null) {
-        return $this->getScenario()->runStep(new \Codeception\Step\Action('assertArrayNotHasKey', func_get_args()));
-    }
-
- 
-    /**
-     * [!] Method is generated. Documentation taken from corresponding module.
-     *
-     * Checks that array contains subset.
-     *
-     * @param array  $subset
-     * @param array  $array
-     * @param bool   $strict
-     * @param string $message
-     * @see \Codeception\Module\Asserts::assertArraySubset()
-     */
-    public function assertArraySubset($subset, $array, $strict = null, $message = null) {
-        return $this->getScenario()->runStep(new \Codeception\Step\Action('assertArraySubset', func_get_args()));
-    }
-
- 
-    /**
-     * [!] Method is generated. Documentation taken from corresponding module.
-     *
-     * @param $expectedCount
-     * @param $actual
-     * @param $description
-     * @see \Codeception\Module\Asserts::assertCount()
-     */
-    public function assertCount($expectedCount, $actual, $description = null) {
-        return $this->getScenario()->runStep(new \Codeception\Step\Action('assertCount', func_get_args()));
-    }
-
- 
-    /**
-     * [!] Method is generated. Documentation taken from corresponding module.
-     *
-     * @param $class
-     * @param $actual
-     * @param $description
-     * @see \Codeception\Module\Asserts::assertInstanceOf()
-     */
-    public function assertInstanceOf($class, $actual, $description = null) {
-        return $this->getScenario()->runStep(new \Codeception\Step\Action('assertInstanceOf', func_get_args()));
-    }
-
- 
-    /**
-     * [!] Method is generated. Documentation taken from corresponding module.
-     *
-     * @param $class
-     * @param $actual
-     * @param $description
-     * @see \Codeception\Module\Asserts::assertNotInstanceOf()
-     */
-    public function assertNotInstanceOf($class, $actual, $description = null) {
-        return $this->getScenario()->runStep(new \Codeception\Step\Action('assertNotInstanceOf', func_get_args()));
-    }
-
- 
-    /**
-     * [!] Method is generated. Documentation taken from corresponding module.
-     *
-     * @param $type
-     * @param $actual
-     * @param $description
-     * @see \Codeception\Module\Asserts::assertInternalType()
-     */
-    public function assertInternalType($type, $actual, $description = null) {
-        return $this->getScenario()->runStep(new \Codeception\Step\Action('assertInternalType', func_get_args()));
-    }
-
- 
-    /**
-     * [!] Method is generated. Documentation taken from corresponding module.
-     *
-     * Fails the test with message.
-     *
-     * @param $message
-     * @see \Codeception\Module\Asserts::fail()
-     */
-    public function fail($message) {
-        return $this->getScenario()->runStep(new \Codeception\Step\Action('fail', func_get_args()));
-    }
-
- 
-    /**
-     * [!] Method is generated. Documentation taken from corresponding module.
-     *
-     * Handles and checks exception called inside callback function.
-     * Either exception class name or exception instance should be provided.
-     *
-     * ```php
-     * <?php
-     * $I->expectException(MyException::class, function() {
-     *     $this->doSomethingBad();
-     * });
-     *
-     * $I->expectException(new MyException(), function() {
-     *     $this->doSomethingBad();
-     * });
-     * ```
-     * If you want to check message or exception code, you can pass them with exception instance:
-     * ```php
-     * <?php
-     * // will check that exception MyException is thrown with "Don't do bad things" message
-     * $I->expectException(new MyException("Don't do bad things"), function() {
-     *     $this->doSomethingBad();
-     * });
-     * ```
-     *
-     * @param $exception string or \Exception
-     * @param $callback
-     *
-     * @deprecated Use expectThrowable instead
-     * @see \Codeception\Module\Asserts::expectException()
-     */
-    public function expectException($exception, $callback) {
-        return $this->getScenario()->runStep(new \Codeception\Step\Action('expectException', func_get_args()));
-    }
-
- 
-    /**
-     * [!] Method is generated. Documentation taken from corresponding module.
-     *
-     * Handles and checks throwables (Exceptions/Errors) called inside the callback function.
-     * Either throwable class name or throwable instance should be provided.
-     *
-     * ```php
-     * <?php
-     * $I->expectThrowable(MyThrowable::class, function() {
-     *     $this->doSomethingBad();
-     * });
-     *
-     * $I->expectThrowable(new MyException(), function() {
-     *     $this->doSomethingBad();
-     * });
-     * ```
-     * If you want to check message or throwable code, you can pass them with throwable instance:
-     * ```php
-     * <?php
-     * // will check that throwable MyError is thrown with "Don't do bad things" message
-     * $I->expectThrowable(new MyError("Don't do bad things"), function() {
-     *     $this->doSomethingBad();
-     * });
-     * ```
-     *
-     * @param $throwable string or \Throwable
-     * @param $callback
-     * @see \Codeception\Module\Asserts::expectThrowable()
-     */
-    public function expectThrowable($throwable, $callback) {
-        return $this->getScenario()->runStep(new \Codeception\Step\Action('expectThrowable', func_get_args()));
-    }
-}
